migrated server.ts into indext.ts and added adzuna controller
