+++ backend/package.json
@@ -7,7 +7,7 @@
     "test": "node src/openAI-test.tsx",
     "build": "tsc",
     "lint": "eslint .",
-    "dev": "nodemon src/index.ts",
+    "dev": "nodemon server.ts",
     "start": "node dist/index.js"
   },
   "keywords": [],
@@ -18,6 +18,7 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "openai": "^4.19.1",
+    "cookie-parser": "^1.4.6",
     "pg": "^8.11.3",
     "webpack-cli": "^5.1.4"
   },
